move browser and reader session bases into subdirs
